+++ src/main/groovy/org/liquibase/gradle/LiquibaseCommand.groovy
@@ -1,0 +1,13 @@
+package org.liquibase.gradle
+
+/**
+ * Created with IntelliJ IDEA.
+ * User: steve
+ * Date: 10/15/22
+ * Time: 1:22 PM
+ * To change this template use File | Settings | File Templates.
+ * 
+ * @author Steven C. Saliman
+ */
+class LiquibaseCommand {
+}