+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+Changes for 2.1.2
+=================
+- Removed the old `liquibase` id from the plugin.  The plugin must be applied with its new standard
+  id of `org.liquibase.gradle`.
+
+- Changed the way Liquibase commands are run. The plugin now sends the correct Liquibase 4.4+
+  commands when it detects a newer Liquibase version on the classpath.  In other words, it will
+  send `drop-all` for Liquibase 4.4+ instead of the legacy `dropAll` command.
+
 Changes for 2.1.1
 =================
 - Fixed the code that auto-detects the version of Liquibase.  Fixed #94